fix(ai): move BaseServer out of shared/ + match Neo namespace convention (#10965)
Per @tobiu's review correction: (1) Neo's class system enforces
filepath-to-className matching, so BaseServer.mjs at
`ai/mcp/server/shared/BaseServer.mjs` cannot use `Neo.mcp.server.Base`.
(2) `shared/` is for cross-cutting *services* (composable primitives),
not class-hierarchy ancestors. BaseServer is an inheritance ancestor;
its natural location is parent-of-server-dirs.

Two changes:
- Moved file: ai/mcp/server/shared/BaseServer.mjs → ai/mcp/server/BaseServer.mjs
- Moved spec: test/playwright/unit/ai/mcp/server/shared/BaseServer.spec.mjs
  → test/playwright/unit/ai/mcp/server/BaseServer.spec.mjs
- className updated: 'Neo.mcp.server.Base' → 'Neo.ai.mcp.server.BaseServer'
- Class JSDoc tag updated to match
- Spec describe-block titles updated to new className for consistency
- Import paths in spec adjusted (shallower by one level)
- TransportService import path adjusted: './services/' → './shared/services/'

26/26 unit tests still pass (842ms).
